add main program and fixed different classes.

+++ pglp_4.2/src/main/java/uvsq2192128/pglp_4_2/SpecificCommand.java
@@ -2,5 +2,5 @@ package uvsq2192128.pglp_4_2;
 
 public interface SpecificCommand extends Command{
 
-	public double apply(int a, int b);
+	public double apply(double a, double b);
 }
